requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
statsmodels
xgboost
lightgbm
python-dotenv
pymongo
requests
certifi
pymongo[srv]==3.12.0
scikit-learn
pyaml==20.4.0
geopy
holidays

#-e .